k8s-api/requirements.txt
celery[redis]<5.1
Django<3.2
djangorestframework<3.13
psycopg2<2.9
requests<2.26
